// packages/connection/src/canisters/nns-dapp.d.ts
export type AccountIdentifier = string;
export interface AttachCanisterRequest {
  name: string;
  canister_id: Principal;
}
export type AttachCanisterResponse =
  | { Ok: null }
  | { CanisterAlreadyAttached: null }
  | { NameAlreadyTaken: null }
  | { NameTooLong: null }
  | { CanisterLimitExceeded: null };
export type BlockHeight = bigint;
// ...
export interface GetTransactionsResponse {
  total: number;
  transactions: Array<Transaction>;
}
// ...
export interface ICPTs {
  e8s: bigint;
}
// ...
export interface Send {
  to: AccountIdentifier;
  fee: ICPTs;
  amount: ICPTs;
}
// ...
export interface Timestamp {
  timestamp_nanos: bigint;
}
export interface Transaction {
  transaction_type: [] | [TransactionType];
  memo: bigint;
  timestamp: Timestamp;
  block_height: BlockHeight;
  transfer: Transfer;
}
export type TransactionType =
  | { Burn: null }
  | { Mint: null }
  | { Send: null }
  | { StakeNeuronNotification: null }
  | { TopUpCanister: CanisterId }
  | { CreateCanister: null }
  | { TopUpNeuron: null }
  | { StakeNeuron: null };
export type Transfer =
  | { Burn: { amount: ICPTs } }
  | { Mint: { amount: ICPTs } }
  | { Send: Send }
  | { Receive: Receive };